Add support for native plugin
Converts the build system to create a container loosely based on
vieux/docker-volume-sshfs and adds in the necessary logic to store/load
the tracked ZFS datasets.

This is a pretty hacky/fragile implementation due to the requirement
that all paths returned by the container be relative to the container's
rootfs path. See https://github.com/moby/moby/pull/26398 for the details
on this. For the specific hack employed here please see the comments on
the constants in `zfs/driver.go`.

Also removes the CLI handling now that we aren't being run by a user.
The debug flag has been replaced with an environment variable that can
be configured on install.